requirements.txt
streamlit==1.53.0
langchain-aws==1.2.1
langgraph==1.0.6
langchain-core==1.2.7
langchain-community==0.4.1
fastmcp==1.25.0
nest-asyncio==1.6.0
python-dotenv==1.2.1
pandas==2.3.3
openpyxl==3.1.5
pypdf==6.6.0
sqlalchemy==2.0.45
langchain-chroma==1.1.0
chromadb==1.4.1